// src/modules/request/dto/search-request.dto.ts
export class SearchPostDto {
  user_id?: string;
  post_type?: string;
  search?: string;
  post_status?: string;
  post_language?: string;
  post_category?: string;
  not_image?: string;
  categories?: any;
  not_download?: string;
  other_status?: string;
  post_parent?: string;
  country?: string;
  from?: string;
  is_pin?: string;
  comment_number?: string;
  channel_id?: string;
  data_json_type?: string;
  ref_id?: string;
}